🚚 refactor(curation): 패키지 구조 정리 및 ViewModel 위치 이동
- **ViewModel 및 네비게이션 경로 재구성**
    - `CurationViewModel` 및 `CurationDetailViewModel`을 `com.linku.curation.viewModel` 패키지로 이동하여 계층 구조 명확화
    - `MainApp` 내 `CurationViewModel` 및 `curationGraph` 임포트 경로를 변경된 패키지 구조에 맞춰 업데이트

- **미사용 코드 제거**
    - UI 모델로 사용되던 `UICurationItem` 데이터 클래스 삭제 (컴포넌트 구조 변경에 따른 정리)
