+++ feature/curation/src/main/java/com/linku/curation/viewModel/CurationViewModel.kt
@@ -1,4 +1,4 @@
-package com.linku.curation
+package com.linku.curation.viewModel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
